Extract blade view methods to a trait
